Use 64 bit build for appveyor

+++ code/graphics/gropengltnl.h
@@ -87,7 +87,7 @@ void gr_opengl_set_team_color(const team_color *colors);
 void opengl_tnl_init();
 void opengl_tnl_shutdown();
 
-void gr_opengl_render_model(model_material* material_info, indexed_vertex_source *vert_source, vertex_buffer* bufferp, int texi);
+void gr_opengl_render_model(model_material* material_info, indexed_vertex_source *vert_source, vertex_buffer* bufferp, size_t texi);
 void opengl_render_model_program(model_material* material_info, indexed_vertex_source *vert_source, vertex_buffer* bufferp, buffer_data *datap);
 void opengl_render_model_fixed(model_material* material_info, indexed_vertex_source *vert_source, vertex_buffer *bufferp, buffer_data *datap);
